[mobile]feat: move home fragment to home folder
